Merging develop into master: July 16th, 2025

+++ Servers/utils/assessment.utils.ts
@@ -1,80 +1,92 @@
-import { Assessment, AssessmentModel } from "../models/assessment.model";
+import { AssessmentModel } from "../domain.layer/models/assessment/assessment.model";
 import { sequelize } from "../database/db";
 import { createNewTopicsQuery } from "./topic.utils";
 import { QueryTypes, Transaction } from "sequelize";
 
-export const getAllAssessmentsQuery = async (): Promise<Assessment[]> => {
+export const getAllAssessmentsQuery = async (
+  tenant: string
+): Promise<AssessmentModel[]> => {
   const assessments = await sequelize.query(
-    "SELECT * FROM assessments ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".assessments ORDER BY created_at DESC, id ASC`,
     {
       mapToModel: true,
-      model: AssessmentModel
+      model: AssessmentModel,
     }
   );
   return assessments;
 };
 
 export const getAssessmentByIdQuery = async (
-  id: number
-): Promise<Assessment | null> => {
+  id: number,
+  tenant: string
+): Promise<AssessmentModel | null> => {
   const result = await sequelize.query(
-    "SELECT * FROM assessments WHERE id = :id",
+    `SELECT * FROM "${tenant}".assessments WHERE id = :id`,
     {
       replacements: { id: id },
       mapToModel: true,
-      model: AssessmentModel
+      model: AssessmentModel,
     }
   );
   return result[0];
 };
 
 export const getAssessmentByProjectIdQuery = async (
-  projectId: number
-): Promise<Assessment[]> => {
+  projectId: number,
+  tenant: string
+): Promise<AssessmentModel[]> => {
   const result = await sequelize.query(
-    "SELECT * FROM assessments WHERE project_id = :project_id ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".assessments WHERE project_id = :project_id ORDER BY created_at DESC, id ASC`,
     {
       replacements: { project_id: projectId },
       mapToModel: true,
-      model: AssessmentModel
+      model: AssessmentModel,
     }
   );
   return result;
 };
 
 export const createNewAssessmentQuery = async (
-  assessment: Assessment,
+  assessment: AssessmentModel,
   enable_ai_data_insertion: boolean,
+  tenant: string,
   transaction: Transaction
 ): Promise<Object> => {
   const result = await sequelize.query(
-    `INSERT INTO assessments (project_id) VALUES (:project_id) RETURNING *`,
+    `INSERT INTO "${tenant}".assessments (project_id) VALUES (:project_id) RETURNING *`,
     {
       replacements: { project_id: assessment.project_id },
       mapToModel: true,
       model: AssessmentModel,
-      transaction
+      transaction,
     }
   );
-  const topics = await createNewTopicsQuery(result[0].id!, enable_ai_data_insertion, transaction);
+  const topics = await createNewTopicsQuery(
+    result[0].id!,
+    enable_ai_data_insertion,
+    tenant,
+    transaction
+  );
   return { assessment: result[0], topics };
 };
 
 export const updateAssessmentByIdQuery = async (
   id: number,
-  assessment: Partial<Assessment>,
+  assessment: Partial<AssessmentModel>,
+  tenant: string,
   transaction: Transaction
 ): Promise<Boolean> => {
   const result = await sequelize.query(
-    `UPDATE assessments SET project_id = :project_id WHERE id = :id RETURNING *`,
+    `UPDATE "${tenant}".assessments SET project_id = :project_id WHERE id = :id RETURNING *`,
     {
       replacements: {
-        project_id: assessment.project_id, id: id
+        project_id: assessment.project_id,
+        id: id,
       },
       mapToModel: true,
       model: AssessmentModel,
       // type: QueryTypes.UPDATE
-      transaction
+      transaction,
     }
   );
   return result.length > 0;
@@ -82,15 +94,16 @@ export const updateAssessmentByIdQuery = async (
 
 export const deleteAssessmentByIdQuery = async (
   id: number,
+  tenant: string,
   transaction: Transaction
 ): Promise<Boolean> => {
   const result = await sequelize.query(
-    `DELETE FROM assessments WHERE id = :id RETURNING *`,
+    `DELETE FROM "${tenant}".assessments WHERE id = :id RETURNING *`,
     {
       replacements: { id: id },
       mapToModel: true,
       model: AssessmentModel,
-      transaction
+      transaction,
     }
   );
   return result.length > 0;
