+++ Servers/utils/control.utils.ts
@@ -1,42 +1,47 @@
-import { Control, ControlModel } from "../models/control.model";
+import { ControlModel } from "../domain.layer/models/control/control.model";
 import { sequelize } from "../database/db";
 import { createNewSubControlsQuery } from "./subControl.utils";
 import { Model, QueryTypes, Transaction } from "sequelize";
+import { IControl } from "../domain.layer/interfaces/i.control";
 
-export const getAllControlsQuery = async (): Promise<Control[]> => {
+export const getAllControlsQuery = async (
+  tenant: string
+): Promise<IControl[]> => {
   const controls = await sequelize.query(
-    "SELECT * FROM controls ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".controls ORDER BY created_at DESC, id ASC`,
     {
       mapToModel: true,
-      model: ControlModel
+      model: ControlModel,
     }
   );
   return controls;
 };
 
 export const getControlByIdQuery = async (
-  id: number
-): Promise<Control | null> => {
+  id: number,
+  tenant: string
+): Promise<IControl | null> => {
   const result = await sequelize.query(
-    "SELECT * FROM controls WHERE id = :id",
+    `SELECT * FROM "${tenant}".controls WHERE id = :id`,
     {
       mapToModel: true,
       model: ControlModel,
-      replacements: { id: id }
+      replacements: { id: id },
     }
   );
   return result[0];
 };
 
 export const getAllControlsByControlGroupQuery = async (
-  controlGroupId: any
-): Promise<Control[]> => {
+  controlGroupId: any,
+  tenant: string
+): Promise<IControl[]> => {
   const controls = await sequelize.query(
-    "SELECT * FROM controls WHERE control_category_id = :control_category_id ORDER BY created_at DESC, id ASC",
+    `SELECT * FROM "${tenant}".controls WHERE control_category_id = :control_category_id ORDER BY created_at DESC, id ASC`,
     {
       replacements: { control_category_id: controlGroupId },
       mapToModel: true,
-      model: ControlModel
+      model: ControlModel,
     }
   );
   return controls;
@@ -46,7 +51,7 @@ export const getControlByIdAndControlTitleAndControlDescriptionQuery = async (
   id: number,
   controlTitle: string,
   controlDescription: string
-): Promise<Control | null> => {
+): Promise<IControl | null> => {
   const result = await sequelize.query(
     `SELECT * FROM controls WHERE
       control_category_id = : control_category_id 
@@ -56,21 +61,22 @@ export const getControlByIdAndControlTitleAndControlDescriptionQuery = async (
       replacements: {
         control_category_id: id,
         control_title: controlTitle,
-        control_description: controlDescription
+        control_description: controlDescription,
       },
       mapToModel: true,
-      model: ControlModel
+      model: ControlModel,
     }
   );
   return result[0];
 };
 
 export const createNewControlQuery = async (
-  control: Partial<Control>,
+  control: Partial<ControlModel>,
+  tenant: string,
   transaction: Transaction
-): Promise<Control> => {
+): Promise<ControlModel> => {
   const result = await sequelize.query(
-    `INSERT INTO controls (
+    `INSERT INTO "${tenant}".controls (
       title, description, order_no, 
       status, approver, risk_review, 
       owner, reviewer, due_date, 
@@ -104,10 +110,11 @@ export const createNewControlQuery = async (
 
 export const updateControlByIdQuery = async (
   id: number,
-  control: Partial<Control>,
+  control: Partial<ControlModel>,
+  tenant: string,
   transaction: Transaction
-): Promise<Control> => {
-  const updateControl: Partial<Record<keyof Control, any>> = {};
+): Promise<ControlModel> => {
+  const updateControl: Partial<Record<keyof ControlModel, any>> = {};
   const setClause = [
     "title",
     "description",
@@ -117,15 +124,22 @@ export const updateControlByIdQuery = async (
     "owner",
     "reviewer",
     "due_date",
-    "implementation_details"
-  ].filter(f => {
-    if (control[f as keyof Control] !== undefined && control[f as keyof Control]) {
-      updateControl[f as keyof Control] = control[f as keyof Control]
-      return true
-    }
-  }).map(f => `${f} = :${f}`).join(", ");
+    "implementation_details",
+  ]
+    .filter((f) => {
+      if (
+        control[f as keyof ControlModel] !== undefined &&
+        control[f as keyof ControlModel]
+      ) {
+        updateControl[f as keyof ControlModel] =
+          control[f as keyof ControlModel];
+        return true;
+      }
+    })
+    .map((f) => `${f} = :${f}`)
+    .join(", ");
 
-  const query = `UPDATE controls SET ${setClause} WHERE id = :id RETURNING *;`;
+  const query = `UPDATE "${tenant}".controls SET ${setClause} WHERE id = :id RETURNING *;`;
 
   updateControl.id = id;
 
@@ -141,10 +155,11 @@ export const updateControlByIdQuery = async (
 
 export const deleteControlByIdQuery = async (
   id: number,
+  tenant: string,
   transaction: Transaction
 ): Promise<Boolean> => {
   const result = await sequelize.query(
-    "DELETE FROM controls WHERE id = :id RETURNING *",
+    `DELETE FROM "${tenant}".controls WHERE id = :id RETURNING *`,
     {
       replacements: { id },
       mapToModel: true,
@@ -173,10 +188,11 @@ export const createNewControlsQuery = async (
     }[];
   }[],
   enable_ai_data_insertion: boolean,
+  tenant: string,
   transaction: Transaction
 ) => {
   const createdControls = [];
-  let query = `INSERT INTO controls(
+  let query = `INSERT INTO "${tenant}".controls(
     title, description, order_no, control_category_id,
     implementation_details, status
   ) VALUES (
@@ -190,8 +206,10 @@ export const createNewControlsQuery = async (
         description: controlStruct.description,
         order_no: controlStruct.order_no,
         control_category_id: controlCategoryId,
-        implementation_details: enable_ai_data_insertion ? controlStruct.implementation_details : null,
-        status: enable_ai_data_insertion ? 'Waiting' : null
+        implementation_details: enable_ai_data_insertion
+          ? controlStruct.implementation_details
+          : null,
+        status: enable_ai_data_insertion ? "Waiting" : null,
       },
       mapToModel: true,
       model: ControlModel,
@@ -203,6 +221,7 @@ export const createNewControlsQuery = async (
       control_id,
       controlStruct.subControls,
       enable_ai_data_insertion,
+      tenant,
       transaction
     );
     createdControls.push({ ...result[0].dataValues, subControls });
